Added the X-Firebase-Decoding: 1 header
When URLs are parsed for the REST API, Firebase decode query parameters twice. Starting February 6, 2019, they will change the behavior so all URLs are parsed correctly (and are RFC compliant, according to RFC 3986, Section 2) .
Until February 6, 2019, we can add the X-Firebase-Decoding: 1 to our requests, to verify if those requests are impacted by this change.